+++ covalent/_results_manager/qresult.py
@@ -23,6 +23,7 @@ from typing import Any
 import pennylane as qml
 from pennylane.tape import QuantumTape
 
+from .._shared_files.qresult_utils import re_execute
 from ..quantum.qclient.core import middleware
 
 
@@ -41,6 +42,7 @@ class QNodeFutureResult:
     def __init__(
         self,
         batch_id: str,
+        interface: str,
         original_qnode: qml.QNode,
         original_tape: QuantumTape,
     ):
@@ -52,14 +54,24 @@ class QNodeFutureResult:
                 the middleware.
         """
         self.batch_id = batch_id
+        self.interface = interface  # NOT the original QNode's interface
 
         # Required for batch_transforms and correct output typing.
         self.device = original_qnode.device
-        self.interface = original_qnode.interface
-        self.diff_method = original_qnode.diff_method
+        self.qnode = original_qnode
         self.tape = original_tape
 
+        self.args = None
+        self.kwargs = None
         self._result = None
+
+    def __call__(self, *args, **kwargs):
+        """
+        Store the arguments and keyword arguments of the original QNode call.
+        """
+        self.args = args
+        self.kwargs = kwargs
+        return self
 
     def result(self) -> Any:
         """
@@ -74,48 +86,29 @@ class QNodeFutureResult:
             # Get raw results from the middleware.
             results = middleware.get_results(self.batch_id)
 
-            # Create a device from the original QNode's device class for correct typing.
-            dev = _run_later_device_factory(results, self.device, self.tape)
+            # Required correct gradient post-processing in some cases.
+            if self.interface == "autograd":
+                self._result = results
+                res = results[0]
 
-            # Define a dummy circuit that returns the original QNode's return value.
-            @qml.qnode(dev, interface=self.interface, diff_method=self.diff_method)
-            def _dummy_circuit():
-                return self.tape._qfunc_output  # pylint: disable=protected-access
+            if self.interface != "numpy":
+                interface = self.interface  # re-execute with any non-numpy interface
+                res = results[0]  # re-execute with this result
 
-            self._result = _dummy_circuit()
+            elif self.qnode.interface is None:
+                interface = None
+                res = results[0]
+
+            elif self.qnode.interface == "auto":
+                interface = "auto"
+                res = results
+
+            else:
+                # Skip re-execution.
+                self._result = results
+                return results
+
+            args, kwargs = self.args, self.kwargs
+            self._result = re_execute(res, self.qnode, self.tape)(interface, *args, **kwargs)
 
         return self._result
-
-
-def _run_later_device_factory(
-    results: Any,
-    original_device: qml.Device,
-    original_tape: QuantumTape,
-) -> qml.Device:
-    """
-    Returns an instance of a new class that inherits from the original QNode's
-    device class. Inheriting ensures the correct return type, while overriding
-    `batch_execute` returns the expected `results` without actually running circuits.
-    """
-
-    qml_device_cls = type(original_device)
-
-    class _RunLaterDevice(qml_device_cls):
-        # pylint: disable=too-few-public-methods
-
-        def batch_execute(self, circuits):
-            """
-            Override to return expected result.
-            """
-            if len(circuits) > 1 or len(results) > 1:
-                return [[r] for r in results]
-            return results
-
-        def batch_transform(self, _):
-            """
-            Ignore blank circuit and run batch transform on original tape.
-            """
-            return original_device.batch_transform(original_tape)
-
-    wires = original_device.num_wires
-    return _RunLaterDevice(wires=wires, shots=1)