+++ covalent/_shared_files/utils.py
@@ -27,9 +27,11 @@ from datetime import timedelta
 from typing import Any, Callable, Dict, Set, Tuple
 
 import cloudpickle
+from pennylane._device import Device
 
 from . import logger
 from .config import get_config
+from .pickling import _qml_mods_pickle
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info
@@ -230,6 +232,7 @@ def get_named_params(func, args, kwargs):
     return (named_args, named_kwargs)
 
 
+@_qml_mods_pickle
 def cloudpickle_serialize(obj):
     return cloudpickle.dumps(obj)
 
@@ -262,3 +265,16 @@ def import_from_path(path: str) -> Any:
     module_path, class_name = path.split(_IMPORT_PATH_SEPARATOR)
     module = importlib.import_module(module_path)
     return getattr(module, class_name)
+
+
+def get_original_shots(dev: Device):
+    """
+    Recreate vector of shots if device has a shot vector.
+    """
+    if not dev.shot_vector:
+        return dev.shots
+
+    shot_sequence = []
+    for shots in dev.shot_vector:
+        shot_sequence.extend([shots.shots] * shots.copies)
+    return type(dev.shot_vector)(shot_sequence)