+++ covalent/_workflow/qdevice.py
@@ -18,10 +18,15 @@
 #
 # Relief from the License may be granted by purchasing a commercial license.
 
+"""
+Define the custom Pennylane device that interacts with Covalent's Quantum Executors.
+"""
+
+from typing import Sequence
+
 from pennylane import QubitDevice
 from pennylane import numpy as np
 from pennylane.devices.default_qubit import DefaultQubit
-from pennylane.gradients import param_shift
 
 from ..quantum.qclient.core import middleware
 
@@ -29,11 +34,14 @@ from ..quantum.qclient.core import middleware
 class QEDevice(QubitDevice):
 
     """
-    Custom Pennylane device that batch executes circuits by submitting them to the middleware
+    The purpose of this device is to redirect circuit execution through Covalent's
+    Quantum Executors and to enable asynchronous execution.
+
+    All QElectrons (which are QNodes) implicitly use this device.
     """
 
     name = "QEDevice"
-    short_name = "qe_device"
+    short_name = "qelectron_device"
     pennylane_requires = ">=0.29.1"
     version = "0.0.1"
     author = "aq"
@@ -59,11 +67,15 @@ class QEDevice(QubitDevice):
         self.executors = executors
         self.qelectron_info = qelectron_info
 
-        # This will be set when the QNodeQE
-        # is called with args and kwargs
+        # This will be set when the QNodeQE is called with args and kwargs.
         self.qnode_specs = None
 
     def batch_execute(self, circuits):
+        """
+        Submits circuits to QElectron middleware for execution on chosen backend(s).
+        Retrieves and returns results, or returns a dummy result if running asynchronously.
+        """
+
         # Async submit all circuits to middleware.
         batch_id = middleware.run_circuits_async(
             circuits, self.executors, self.qelectron_info, self.qnode_specs
@@ -73,63 +85,18 @@ class QEDevice(QubitDevice):
         # We will retrieve the result later using the future object and the batch_id.
         if self._async_run:
             self._batch_id = batch_id
-            # Return a dummy result
-            return [self._asarray([1] * c.output_dim) for c in circuits]
+
+            # Return a (recognizable) dummy result
+            res = [self._asarray([-123.456] * c.output_dim) for c in circuits]
+            if isinstance(self.qelectron_info.device_shots, Sequence):
+                # Replicate the list for shot vector case.
+                return [[res]] * len(self.qelectron_info.device_shots)
+            return res
 
         # Otherwise, get the results from the middleware
         results = middleware.get_results(batch_id)
 
         return results
-
-    def gradients(self, circuits, method="jacobian", **kwargs):
-        # Obtain gradient tapes and processing functions.
-        grad_tapes_tuple, processing_fns = zip(
-            *(param_shift(circuit, **kwargs) for circuit in circuits)
-        )
-
-        # Flatten the list of gradient tapes.
-        nonempty_grad_tapes_list = []
-        for grad_tapes in grad_tapes_tuple:
-            nonempty_grad_tapes_list.extend(grad_tapes)
-
-        # Skip gradient computation for vectorized calls.
-        # Also require nonempty list of gradient tapes.
-        scalar_input = all(not circuit.batch_size for circuit in circuits)
-        if not (scalar_input and nonempty_grad_tapes_list):
-            return []
-
-        # Submit gradient circuits to middleware.
-        nonempty_grad_res = self.batch_execute(nonempty_grad_tapes_list)
-
-        # Unpack the nonempty gradient result into the expected shape.
-        grad_res = self._unpack_gradient_result(nonempty_grad_res, grad_tapes_tuple)
-
-        # Compute parameter-shift jacobians.
-        jacs = [fn(r) for r, fn in zip(grad_res, processing_fns)]
-
-        return jacs
-
-    def _unpack_gradient_result(self, nonzero_grads, grad_tapes_tuple):
-        """
-        Unpack the gradient tapes' execution results into a list containing arrays
-        and lists of arrays. Insert zero-arrays for empty gradient tapes.
-        """
-        grad_res = []
-        zero_arr = self._asarray(0.0)
-        nonempty_idx = 0
-        for grad_tapes in grad_tapes_tuple:
-            if not grad_tapes:
-                grad_res.append(zero_arr)
-            else:
-                sub_res = []
-                for _ in grad_tapes:
-                    sub_res.append(nonzero_grads[nonempty_idx])
-                    nonempty_idx += 1
-
-                sub_res = sub_res[0] if len(sub_res) == 1 else sub_res
-                grad_res.append(sub_res)
-
-        return grad_res
 
     def apply(self, *args, **kwargs):
         # Dummy implementation of abstractmethod on `QubitDevice`