+++ covalent/executor/quantum_plugins/aws_plugin/aws_plugin.py
@@ -92,7 +92,6 @@ class BraketQubitExecutor(BaseThreadPoolQExecutor):
         default_factory=lambda: get_config("qelectron")["BraketQubitExecutor"]["max_retries"]
     )
     max_jobs: int = 20
-    shots: Optional[int] = 0
     aws_session: Optional[str] = None  # not actually a str. Fix.
     parallel: bool = False
     max_parallel: Optional[int] = None
@@ -110,18 +109,15 @@ class BraketQubitExecutor(BaseThreadPoolQExecutor):
             jobs: a :code:`list` of tasks subitted by threads.
         """
 
-        # Check `self.shots` against 0 to allow override with `None`.
-        device_shots = self.shots if self.shots != 0 else self.qnode_device_shots
-
         p = get_thread_pool(self.max_jobs)
         jobs = []
         for qscript in qscripts_list:
             dev = qml.device(
                 "braket.aws.qubit",
-                wires=qscript.wires,
+                wires=self.qelectron_info.device_wires,
                 device_arn=self.device_arn,
                 s3_destination_folder=self.s3_destination_folder,
-                shots=device_shots,
+                shots=self.override_shots,
                 poll_timeout_seconds=self.poll_timeout_seconds,
                 poll_interval_seconds=self.poll_interval_seconds,
                 aws_session=self.aws_session,
@@ -165,7 +161,6 @@ class LocalBraketQubitExecutor(BaseThreadPoolQExecutor):
         default_factory=lambda: get_config("qelectron")["LocalBraketQubitExecutor"]["backend"]
     )
     max_jobs: int = 20
-    shots: Optional[int] = 0
     run_kwargs: dict = {}
 
     def batch_submit(self, qscripts_list):
@@ -179,17 +174,14 @@ class LocalBraketQubitExecutor(BaseThreadPoolQExecutor):
             jobs: a :code:`list` of tasks subitted by threads.
         """
 
-        # Check `self.shots` against 0 to allow override with `None`.
-        device_shots = self.shots if self.shots != 0 else self.qnode_device_shots
-
         p = get_thread_pool(self.max_jobs)
         jobs = []
         for qscript in qscripts_list:
             dev = qml.device(
                 "braket.local.qubit",
-                wires=qscript.wires,
+                wires=self.qelectron_info.device_wires,
                 backend=self.backend,
-                shots=device_shots,
+                shots=self.override_shots,
                 **self.run_kwargs,
             )