+++ covalent/executor/quantum_plugins/qiskit_plugin/runtime_sampler.py
@@ -22,13 +22,13 @@
 Pennylane-Qiskit device that uses the Qiskit Runtime `Sampler` primitive
 """
 
-from typing import Any, List, Union
+from typing import Any, List, Sequence, Union
 
 import pennylane as qml
 from devices_base import QiskitSamplerDevice
 from qiskit_ibm_runtime import Sampler
+from qiskit_utils import extract_options
 from sessions import get_cached_session
-from utils import extract_options
 
 
 class QiskitRuntimeSampler(QiskitSamplerDevice):
@@ -41,7 +41,7 @@ class QiskitRuntimeSampler(QiskitSamplerDevice):
     def __init__(  # pylint: disable=too-many-arguments
         self,
         wires: int,
-        shots: int,
+        shots: Union[None, int, Sequence[int], Sequence[Union[int, Sequence[int]]]],
         backend_name: str,
         local_transpile: bool,
         max_time: Union[int, str],
@@ -216,7 +216,7 @@ class QiskitRuntimeSampler(QiskitSamplerDevice):
         as if it came from `batch_execute`.
         """
         dev = _PostProcessDevice(self.wires, results)
-        return qml.execute(tapes, dev, None)
+        return qml.execute(tapes, dev, gradient_fn="best")
 
     def _dummy_result(self) -> Union[Any, List[Any]]:
         """