+++ covalent_ui/webapp/src/components/common/QElectronDrawer.js
@@ -107,6 +107,9 @@ const QElectronDrawer = ({ toggleQelectron, openQelectronDrawer, dispatchId, ele
 
   useEffect(() => {
     setDefaultId('');
+    if (openQelectronDrawer) {
+      setExpanded(true);
+    }
     if (!(electronId === null || electronId === undefined) && openQelectronDrawer) {
       const bodyParams = {
         sort_by: 'start_time',
@@ -145,6 +148,7 @@ const QElectronDrawer = ({ toggleQelectron, openQelectronDrawer, dispatchId, ele
               cursor: 'pointer',
             }}
             component={closeIcon}
+            data-testid="qElectronDrawerSnackbar"
             onClick={() => setOpenSnackbar(false)}
           />
         }
@@ -177,7 +181,7 @@ const QElectronDrawer = ({ toggleQelectron, openQelectronDrawer, dispatchId, ele
         variant="persistent"
         open={openQelectronDrawer}
         onClose={handleDrawerClose}
-        data-testid="nodeDrawer"
+        data-testid="qElectronDrawer"
       >
         <Grid
           container
@@ -188,7 +192,7 @@ const QElectronDrawer = ({ toggleQelectron, openQelectronDrawer, dispatchId, ele
               details={details}
               toggleQelectron={toggleQelectron}
             />
-            <QElelctronAccordion expanded={expanded} setExpanded={setExpanded} overviewData={overviewData} />
+            <QElelctronAccordion expanded={expanded} setExpanded={setExpanded} overviewData={overviewData} openQelectronDrawer={openQelectronDrawer} />
             <QElectronList expanded={expanded}
               data={listData}
               rowClick={rowClickHandler}