+++ covalent_ui/webapp/src/components/qelectron/Overview.js
@@ -48,7 +48,7 @@ const Overview = (props) => {
       >
         Execution Details
       </Typography>
-      <Grid container px={2} py={1} direcction="row">
+      <Grid container px={2} py={1} direcction="row" data-testid="Overview-grid">
         <Grid id="leftGrid" item xs={6}>
           <Typography
             sx={{